Cleaned up file names
