+++ .gitignore
@@ -1,6 +1,9 @@
 # PyCharm files
 .idea/
 
+#VS Code files
+.vscode/
+
 # Mac Temp Files
 .DS_Store
