+++ autonomy_main_Color.py
@@ -10,13 +10,13 @@ from drivers.rovecomm import RoveCommEthernetUdp
 from drivers.drive_board import DriveBoard
 from drivers.nav_board import NavBoard
 # from algorithms.objecttracking import ObjectTracker
-from algorithms.ColorBasedTracking import ObjectTracker
+from algorithms.color_based_tracking import ObjectTracker
 # from algorithms.CannyTracking import ObjectTracker
 import algorithms.gps_navigate as gps_nav
 import algorithms.marker_search as marker_search
 from algorithms.gps_navigate import GPSData
 import algorithms.geomath as geomath
-import algorithms.followBall as follow_ball
+import algorithms.follow_ball 
 
 # Hardware Setup
 rovecomm_node = RoveCommEthernetUdp()
