+++ tests/mag_calibration.py
