+++ tests/record_cameras.py
